chore: :truck: page 폴더 이동 (app 라우터와 겹치지 않도록)
